% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Drop-in replacement for \link[dplyr:mutate]{dplyr::mutate}}
\usage{
mutate(x, ..., .by, .order_by, .frame, .complete = FALSE)
}
\arguments{
\item{x}{(\code{data.frame}_}

\item{...}{expressions to be passed to \link[dplyr:mutate]{dplyr::mutate}}

\item{.by}{(expression, optional: Yes) Columns to group by}

\item{.order_by}{(expression, optional: Yes) Columns to order by}

\item{.frame}{(vector, optional: Yes) Object of class \code{frame} created by one
of these functions: \code{rows_between}, \code{range_between}}

\item{.complete}{(flag, default: FALSE) passed to \link[slider:slide]{slider::slide}}
}
\value{
\code{data.frame}
}
\description{
Provides supercharged version of \link[dplyr:mutate]{dplyr::mutate}
with \code{.by} (group by), \code{.order_by} and \code{.frame} aggregation over arbitrary
window frame
}
\details{
A window function returns a value for every input row of a dataframe
based on a group of rows (frame) in the neighborhood of the input row. This
function implements computation over groups (\code{partition_by} in SQL) in a
predefined order (\code{order_by} in SQL) across a neighborhood of rows (frame)
defined by
\itemize{
\item \code{rows_between}: Number of rows before and after the corresponding row.
Example: \code{c(2, 1)}
\item \code{range_between}: Range is numeric or interval objects
Example: \code{c(days(2), days(1))}
}

This implementation is inspired by spark's \href{https://www.databricks.com/blog/2015/07/15/introducing-window-functions-in-spark-sql.html}{window API}.
The output has the same row order as the input independent of the
\code{order_by}.
}
\examples{
library("magrittr") # for pipe
# example 1: rows between
# Using iris dataset,
# compute cumulative mean of column `Sepal.Length`
# ordered by `Petal.Width` and `Sepal.Width` columns
# grouped by `Petal.Length` column

iris \%>\%
  mutate(sl_mean = mean(Sepal.Length),
         .order_by = c(Petal.Width, Sepal.Width),
         .by = Petal.Length,
         .frame = rows_between(Inf, 0),
         ) \%>\%
  dplyr::slice_min(n = 3, Petal.Width, by = Species)

# example 2: range between
# Using a sample airquality dataset,
# compute mean temp over last seven days in the same month for every row

set.seed(101)
airquality \%>\%
  # create date column
  dplyr::mutate(date_col = lubridate::make_date(1973, Month, Day)) \%>\%
  # create gaps by removing some days
  dplyr::slice_sample(prop = 0.8) \%>\%
  # compute mean temperature over last seven days in the same month
  tidier::mutate(avg_temp_over_last_week = mean(Temp, na.rm = TRUE),
                 .order_by = date_col,
                 .by = Month,
                 .frame = range_between(
                            lubridate::days(7), # 7 days before current row
                            lubridate::days(-1) # do not include current row
                            )
                 )

# example 3: custom function / modeling over window frame
fit_lm_safe = function(df, formula, min_rows = 2) {
  if (is.null(df) || nrow(df) < min_rows) {
    return(NULL)
  }

  tryCatch(
    lm(formula, data = df),
    error = function(e) NULL
  )
}

mtcars \%>\%
  mutate(s = list(fit_lm_safe(pick(everything()), mpg ~ .)),
         .by = c(cyl, vs),
         .order_by = qsec,
         .frame = range_between(-1, 3),
         .complete = FALSE
         ) \%>\%
  head(10)

\dontrun{
# example 4: parallel execution across many groups using tidyr::nest and furrr
future::plan(future::multisession(workers = 2))

iris \%>\%
  tidyr::nest(.by = Species) \%>\%
  dplyr::mutate(
    data = furrr::future_map(data, ~ .x \%>\%
                        mutate(
                          sl_mean = mean(Sepal.Length),
                          .order_by = Petal.Width,
                          .frame = rows_between(2, 2)
                        )
                     )
  ) \%>\%
  tidyr::unnest(data)
}
}
\seealso{
\code{\link[=rows_between]{rows_between()}}, \code{\link[=range_between]{range_between()}}, \code{\link[=mutate]{mutate()}}
}
