% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{range_between}
\alias{range_between}
\title{Create a frame indicating range between window}
\usage{
range_between(before, after)
}
\arguments{
\item{before}{range of rows before}

\item{after}{range of rows after}
}
\value{
object of class range_between_frame, frame
}
\description{
Create a frame indicating range between window
}
\seealso{
\code{\link[=rows_between]{rows_between()}}, \code{\link[=range_between]{range_between()}}, \code{\link[=mutate]{mutate()}}
}
