% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{remove_common_nested_columns}
\alias{remove_common_nested_columns}
\title{Remove non-list columns when same are present in a list column}
\usage{
remove_common_nested_columns(df, list_column)
}
\arguments{
\item{df}{input dataframe}

\item{list_column}{Name or expr of the column which is a list of named lists}
}
\value{
dataframe
}
\description{
Remove non-list columns when same are present in a list column
}
\keyword{internal}
