% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{rows_between}
\alias{rows_between}
\title{Create a frame indicating rows between window}
\usage{
rows_between(before, after)
}
\arguments{
\item{before}{Number of rows before}

\item{after}{Number of rows after}
}
\value{
object of class rows_between_frame, frame
}
\description{
Create a frame indicating rows between window
}
\seealso{
\code{\link[=rows_between]{rows_between()}}, \code{\link[=range_between]{range_between()}}, \code{\link[=mutate]{mutate()}}
}
